+++ packages/amplify_flutter/lib/categories/amplify_analytics_category.dart
@@ -20,9 +20,22 @@ class AnalyticsCategory {
 
   static List<AnalyticsPluginInterface> plugins = [];
 
-  bool addPlugin(AnalyticsPluginInterface plugin) {
-    plugins.add(plugin);
-    return true;
+  Future<void> addPlugin(AnalyticsPluginInterface plugin) async {
+    //TODO: Allow for multiple plugins to work simultaneously
+    if (plugins.length == 0) {
+      try {
+        await plugin.addPlugin();
+        plugins.add(plugin);
+      } on AmplifyAlreadyConfiguredException catch (e) {
+        plugins.add(plugin);
+      } on PlatformException catch (e) {
+        throw AmplifyException.fromMap(
+            Map<String, String>.from(e.details));
+      }
+    } else {
+      throw AmplifyException("Analytics plugin has already been added, " +
+          "multiple plugins for Analytics category are currently not supported.");
+    }
   }
 
   Future<void> recordEvent({AnalyticsEvent event}) async {