+++ packages/amplify_flutter/lib/categories/amplify_api_category.dart
@@ -20,12 +20,18 @@ class APICategory {
 
   static List<APIPluginInterface> plugins = [];
 
-  /// `Add plugin` method
-  bool addPlugin(APIPluginInterface plugin) {
+  Future<void> addPlugin(APIPluginInterface plugin) async {
     //TODO: Allow for multiple plugins to work simultaneously
     if (plugins.length == 0) {
-      plugins.add(plugin);
-      return true;
+      try {
+        await plugin.addPlugin();
+        plugins.add(plugin);
+      } on AmplifyAlreadyConfiguredException catch (e) {
+        plugins.add(plugin);
+      } on PlatformException catch (e) {
+        throw AmplifyException.fromMap(
+            Map<String, String>.from(e.details));
+      }
     } else {
       throw AmplifyException("API plugin has already been added, " +
           "multiple plugins for API category are currently not supported.");