+++ packages/amplify_storage_s3/ios/Classes/SwiftAmplifyStorageS3Plugin.swift
@@ -18,36 +18,72 @@ import UIKit
 import Amplify
 import AmplifyPlugins
 import AWSMobileClient
+import amplify_core
 
 public class SwiftAmplifyStorageS3Plugin: NSObject, FlutterPlugin {
     public static func register(with registrar: FlutterPluginRegistrar) {
         let channel = FlutterMethodChannel(name: "com.amazonaws.amplify/storage_s3", binaryMessenger: registrar.messenger())
         let instance = SwiftAmplifyStorageS3Plugin()
         registrar.addMethodCallDelegate(instance, channel: channel)
-        let storagePlugin = AWSS3StoragePlugin()
-        do {
-            try Amplify.add(plugin: storagePlugin)
-        } catch {
-            print ("Failed to add Amplify AWSS3StoragePlugin \(error)")
-        }
     }
 
   public func handle(_ call: FlutterMethodCall, result: @escaping FlutterResult) {
     print("In handle for method \(call.method)")
+
+    if(call.method == "addPlugin"){
+        do {
+            try Amplify.add(plugin: AWSS3StoragePlugin() )
+            result(true)
+        } catch let error{
+            if(error is StorageError){
+                let storageError = error as! StorageError
+
+                ErrorUtil.postErrorToFlutterChannel(
+                    result: result,
+                    errorCode: "StorageException",
+                    details: [
+                        "message" : storageError.errorDescription,
+                        "recoverySuggestion" : storageError.recoverySuggestion,
+                        "underlyingError": storageError.underlyingError != nil ? storageError.underlyingError!.localizedDescription : ""
+                    ])
+            } else if(error is ConfigurationError) {
+                let configError = error as! ConfigurationError
+                
+                var errorCode = "StorageException"
+                if case .amplifyAlreadyConfigured = configError {
+                   errorCode = "AmplifyAlreadyConfiguredException"
+                }
+                ErrorUtil.postErrorToFlutterChannel(
+                    result: result,
+                    errorCode: errorCode,
+                    details: [
+                        "message" : configError.errorDescription,
+                        "recoverySuggestion" : configError.recoverySuggestion,
+                        "underlyingError": configError.underlyingError != nil ? configError.underlyingError!.localizedDescription : ""
+                    ]
+                )
+            } else{
+                print("Failed to add Amplify Storage Plugin \(error)")
+                result(false)
+            }
+        }
+        return
+    }
+
     let arguments = call.arguments as! Dictionary<String,AnyObject>
     switch call.method {
-    case "uploadFile":
-        AmplifyStorageOperations.uploadFile(flutterResult: result, request: arguments)
-    case "getUrl":
-        AmplifyStorageOperations.getURL(flutterResult: result, request: arguments)
-    case "remove":
-        AmplifyStorageOperations.remove(flutterResult: result, request: arguments)
-    case "list":
-        AmplifyStorageOperations.list(flutterResult: result, request: arguments)
-    case "downloadFile":
-        AmplifyStorageOperations.downloadFile(flutterResult: result, request: arguments)
-    default:
-        result(FlutterMethodNotImplemented)
-     }
-}
+        case "uploadFile":
+            AmplifyStorageOperations.uploadFile(flutterResult: result, request: arguments)
+        case "getUrl":
+            AmplifyStorageOperations.getURL(flutterResult: result, request: arguments)
+        case "remove":
+            AmplifyStorageOperations.remove(flutterResult: result, request: arguments)
+        case "list":
+            AmplifyStorageOperations.list(flutterResult: result, request: arguments)
+        case "downloadFile":
+            AmplifyStorageOperations.downloadFile(flutterResult: result, request: arguments)
+        default:
+            result(FlutterMethodNotImplemented)
+    }
+  }
 }