+++ packages/amplify_flutter/CHANGELOG.md
@@ -1,3 +1,29 @@
+## 0.1.4 (2021-04-27)
+
+### Bug Fixes
+
+- fix: swallows exception on hot restart (#527)
+
+## 0.1.3 (2021-04-21)
+
+## 0.1.2 (2021-04-16)
+
+### Bug Fixes
+
+- fix: handle hot restart (#491)
+
+## 0.1.1 (2021-03-29)
+
+### Chores
+
+- chore: remove upper constraints for flutter 2.0 (#479)
+- chore(amplify_flutter): Add isConfigured API to Amplify (#452)
+
+### Bug Fixes
+
+- fix: Move AddPlugin from Register to MethodChannel (#411)
+- fix: upgrade plugin_platform_interface (#447)
+
 ## 0.1.0 (2021-02-15)
 
 ### Chores